+++ tests/test_E4_verification.py
@@ -1,4 +1,4 @@
-import mvs_eland.E4_verification_of_constraints as E4
+import mvs_eland.E4_verification as E4
 
 from mvs_eland.utils.constants_json_strings import (
     CONSTRAINTS,