+++ CHANGELOG.md
@@ -25,6 +25,10 @@ Here is a template for new release sections
 - Instruction to install graphviz on windows in `docs/troubleshooting.rst` (#572)
 - Benchmark test `test_benchmark_feature_parameters_as_timeseries` to ensure that parameters can always also be defined as a timeseries. Applied to `efficiency` of an energyConversion asset and `electricity_price` of an energyProduction asset (#542)
 - Input files for benchmark tests `test_benchmark_feature_input_flows_as_list` (`Feature_input_flows_as_list`) and `test_benchmark_feature_output_flows_as_list` (`Feature_output_flows_as_list`), but not the benchmark assertions (#542)
+- Error message if time series of non-dispatchable sources do not meet requirement: values betw. 0 and 1. (#498)
+- Requirement for time series of non-dispatchable sources in readthedocs (#498)
+- Provide a warning in case of excessive excess generation (#498)
+- Pytests for `C0.add_maximum_cap()`, renamed function into `C0.process_maximum_cap_constraint()` (#498)
 
 ### Changed
 - Modify `setup.py` to upload the code as package on pypi.org (#570)
@@ -33,6 +37,7 @@ Here is a template for new release sections
 - Update CONTRIBUTING to add a "Write test for your code" section before the "Run tests locally" one (#579)
 - Modified readthedocs page describing the parameters of MVS (#479)
 - Changed `E2.calculate_dispatch_expenditures()` so that it can process parameters defined as lists (#542)
+- Rename `E4` to `E4_verification.py` (#498)
 
 ### Removed
 -
@@ -40,6 +45,9 @@ Here is a template for new release sections
 ### Fixed
 - Update the release protocol in `CONTRIBUTING.md` file (#576)
 - Fix reading timeseries for parameters in `C0` (#542)
+- Constraint for `optimizedAddCap` of non-dispatchable sources: multiply `maximumCap` by `max(timeseries(kWh/kWp))` to fix issue #446 (#562, #498)
+-`timeseries_normalized` are calculated for all `timeseries` of non-dispatchable sources now (before only if `optimizeCap==True`) (#562, #498)
+- Input files of benchmark test `Test_Constraints.test_benchmark_minimal_renewable_share_constraint()` (#498)
 
 ## [0.4.1] - 2020-09-21